Use scan-build in automated CI
Unfortunately, for the time being there is no way to
make the build matrix not go crazy, so the current infrastructure
results in 6 builds :( It's not the end of the world though, since
this project isn't as busy as it could be. Testing to see how horrible
this is now.