Implemented WebVTT document parser, and additional string helpers
String helpers include parseInt(), which parses an integer from a
given position with a radix of 10 (Perhaps a future version may be
added later which is more flexible).

Additionally, another one will skip HTML5 space characters, and
there is additionally a new substring method, however it is likely
to be broken due to refcounting issues, and will need some testing.

Unit tests for these new methods, including the document parsing,
are forthcoming.

Note that the document parser is still not able to actually dispatch
Cues to the client, as cues have not been properly defined as of yet.
Further, there is still no support for dispatching errors to serve
linters/validators. Sorry!