Add tests for setting WebVTT attributes on EmptyCues
Tests to check that setAlign/Line/Position/Size/Vertical fail on
EmptyCues, even when the values are "okay". Blues a bunch of orange
lines in the lcov reports :>

Improved test coverage provides a test case to ensure that EmptyCues
cannot have WebVTT properties set. This is going to be re-structured
soon, but for now it brings Cue coverage to nearly 100%.