Added more Node tests and fixed some bugs
The Node API has changed slightly (the children() method
will only return a const reference to the internal node list).
This shouldn't be a problem because using the children() method
isn't very useful anyways, so don't do it!

There are new tests for the shared InternalNode operations, as well
as some tests for the VoiceNode. Tests for other WebVTT internal
nodes are forthcoming.