Add test for nested ruby-text markup
This is currently the most complicated test, and a few things
needed to be fixed in order to work correctly. I'm not sure if the
way things were fixed is necessarily correct, hoever. Prbably the
right thing to do is make some changes to WebVTTTokenizer::nextChar(),
rather than support byte-by-byte appends