chore(LICENSE): remove silly license restrictions
I was upset with Mozilla, and with Seneca College. I fully admit this. But, what the heck, I admittedly don't give a flying fuck about any of this, at this point. So fuck it, anyone who wants to can feel free to do whatever they want with this.

Admittedly, I haven't been maintaining it for a while, and it's definitely fallen out of sync with the specification, so it would need some legwork to bring it back.